+++ lib/src/flexible_bottom_sheet_route.dart
@@ -43,6 +43,7 @@ const Duration _bottomSheetDuration = Duration(milliseconds: 500);
 /// [duration] - animation speed when opening bottom sheet.
 /// [isSafeArea] - should the bottom sheet provide a SafeArea, false by default.
 /// [decoration] - content decoration bottom sheet.
+/// [useRootScaffold] - if true, add Scaffold widget on widget tree. Default true.
 Future<T?> showFlexibleBottomSheet<T>({
   required BuildContext context,
   required FlexibleDraggableScrollableWidgetBuilder builder,
@@ -63,6 +64,7 @@ Future<T?> showFlexibleBottomSheet<T>({
   Duration? duration,
   bool isSafeArea = false,
   BoxDecoration? decoration,
+  bool useRootScaffold = true,
 }) {
   assert(debugCheckHasMediaQuery(context));
   assert(debugCheckHasMaterialLocalizations(context));
@@ -89,6 +91,7 @@ Future<T?> showFlexibleBottomSheet<T>({
       duration: duration,
       isSafeArea: isSafeArea,
       decoration: decoration,
+      useRootScaffold: useRootScaffold,
     ),
   );
 }
@@ -123,6 +126,7 @@ Future<T?> showFlexibleBottomSheet<T>({
 /// [barrierColor] - barrier color, if you pass [barrierColor] - [isModal] must be true.
 /// [duration] - animation speed when opening bottom sheet.
 /// [isSafeArea] - should the bottom sheet provide a SafeArea, false by default.
+/// [useRootScaffold] - if true, add Scaffold widget on widget tree. Default true.
 Future<T?> showStickyFlexibleBottomSheet<T>({
   required BuildContext context,
   required FlexibleDraggableScrollableHeaderWidgetBuilder headerBuilder,
@@ -147,6 +151,7 @@ Future<T?> showStickyFlexibleBottomSheet<T>({
   Color? barrierColor,
   Duration? duration,
   bool isSafeArea = false,
+  bool useRootScaffold = true,
 }) {
   assert(maxHeaderHeight != null || headerHeight != null);
   assert(debugCheckHasMediaQuery(context));
@@ -177,6 +182,7 @@ Future<T?> showStickyFlexibleBottomSheet<T>({
       barrierBottomSheetColor: barrierColor,
       duration: duration,
       isSafeArea: isSafeArea,
+      useRootScaffold: useRootScaffold,
     ),
   );
 }
@@ -205,6 +211,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
   final Color? barrierBottomSheetColor;
   final Duration? duration;
   final bool isSafeArea;
+  final bool useRootScaffold;
 
   @override
   final String? barrierLabel;
@@ -231,6 +238,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
     required this.isExpand,
     required this.isModal,
     required this.isSafeArea,
+    required this.useRootScaffold,
     this.draggableScrollableController,
     this.builder,
     this.headerBuilder,
@@ -285,6 +293,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
               decoration: decoration,
               keyboardBarrierColor: keyboardBarrierColor,
               bottomSheetColor: bottomSheetColor,
+              useRootScaffold: useRootScaffold,
               bottomSheetBorderRadius: bottomSheetBorderRadius,
             )
           : FlexibleBottomSheet(
@@ -303,6 +312,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
               decoration: decoration,
               keyboardBarrierColor: keyboardBarrierColor,
               bottomSheetColor: bottomSheetColor,
+              useRootScaffold: useRootScaffold,
               bottomSheetBorderRadius: bottomSheetBorderRadius,
             ),
     );