+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased, estimated N.N.N]
+## 3.2.0
+### Added
+* Added a new parameter that controls the display of the snackbar
+### Fixed
+* Added Force widget rebuild to change bottomSheetOffset value
 
 ## 3.1.2
 ### Fixed